+++ sdk/recoveryservices/azure-mgmt-recoveryservicesbackup/azure/mgmt/recoveryservicesbackup/models/_models.py
@@ -742,11 +742,15 @@ class AzureFileshareProtectedItem(ProtectedItem):
      'ProtectionStopped', 'ProtectionPaused'
     :type protection_state: str or
      ~azure.mgmt.recoveryservicesbackup.models.ProtectionState
-    :param health_status: backups running status for this backup item.
-     Possible values include: 'Passed', 'ActionRequired', 'ActionSuggested',
-     'Invalid'
+    :param health_status: Health status of protected item. Possible values
+     include: 'Passed', 'ActionRequired', 'ActionSuggested', 'Healthy',
+     'TransientDegraded', 'PersistentDegraded', 'TransientUnhealthy',
+     'PersistentUnhealthy', 'Invalid'
     :type health_status: str or
      ~azure.mgmt.recoveryservicesbackup.models.HealthStatus
+    :param health_details: Health details on this backup item.
+    :type health_details:
+     list[~azure.mgmt.recoveryservicesbackup.models.HealthDetails]
     :param last_backup_status: Last backup operation status. Possible values:
      Healthy, Unhealthy.
     :type last_backup_status: str
@@ -781,6 +785,7 @@ class AzureFileshareProtectedItem(ProtectedItem):
         'protection_status': {'key': 'protectionStatus', 'type': 'str'},
         'protection_state': {'key': 'protectionState', 'type': 'str'},
         'health_status': {'key': 'healthStatus', 'type': 'str'},
+        'health_details': {'key': 'healthDetails', 'type': '[HealthDetails]'},
         'last_backup_status': {'key': 'lastBackupStatus', 'type': 'str'},
         'last_backup_time': {'key': 'lastBackupTime', 'type': 'iso-8601'},
         'extended_info': {'key': 'extendedInfo', 'type': 'AzureFileshareProtectedItemExtendedInfo'},
@@ -792,6 +797,7 @@ class AzureFileshareProtectedItem(ProtectedItem):
         self.protection_status = kwargs.get('protection_status', None)
         self.protection_state = kwargs.get('protection_state', None)
         self.health_status = kwargs.get('health_status', None)
+        self.health_details = kwargs.get('health_details', None)
         self.last_backup_status = kwargs.get('last_backup_status', None)
         self.last_backup_time = kwargs.get('last_backup_time', None)
         self.extended_info = kwargs.get('extended_info', None)
@@ -1429,7 +1435,9 @@ class AzureIaaSVMProtectedItem(ProtectedItem):
     :type protection_state: str or
      ~azure.mgmt.recoveryservicesbackup.models.ProtectionState
     :param health_status: Health status of protected item. Possible values
-     include: 'Passed', 'ActionRequired', 'ActionSuggested', 'Invalid'
+     include: 'Passed', 'ActionRequired', 'ActionSuggested', 'Healthy',
+     'TransientDegraded', 'PersistentDegraded', 'TransientUnhealthy',
+     'PersistentUnhealthy', 'Invalid'
     :type health_status: str or
      ~azure.mgmt.recoveryservicesbackup.models.HealthStatus
     :param health_details: Health details on this backup item.
@@ -1567,7 +1575,9 @@ class AzureIaaSClassicComputeVMProtectedItem(AzureIaaSVMProtectedItem):
     :type protection_state: str or
      ~azure.mgmt.recoveryservicesbackup.models.ProtectionState
     :param health_status: Health status of protected item. Possible values
-     include: 'Passed', 'ActionRequired', 'ActionSuggested', 'Invalid'
+     include: 'Passed', 'ActionRequired', 'ActionSuggested', 'Healthy',
+     'TransientDegraded', 'PersistentDegraded', 'TransientUnhealthy',
+     'PersistentUnhealthy', 'Invalid'
     :type health_status: str or
      ~azure.mgmt.recoveryservicesbackup.models.HealthStatus
     :param health_details: Health details on this backup item.
@@ -1784,7 +1794,9 @@ class AzureIaaSComputeVMProtectedItem(AzureIaaSVMProtectedItem):
     :type protection_state: str or
      ~azure.mgmt.recoveryservicesbackup.models.ProtectionState
     :param health_status: Health status of protected item. Possible values
-     include: 'Passed', 'ActionRequired', 'ActionSuggested', 'Invalid'
+     include: 'Passed', 'ActionRequired', 'ActionSuggested', 'Healthy',
+     'TransientDegraded', 'PersistentDegraded', 'TransientUnhealthy',
+     'PersistentUnhealthy', 'Invalid'
     :type health_status: str or
      ~azure.mgmt.recoveryservicesbackup.models.HealthStatus
     :param health_details: Health details on this backup item.
@@ -1882,7 +1894,45 @@ class AzureIaaSVMErrorInfo(Model):
         self.recommendations = None
 
 
-class AzureIaaSVMHealthDetails(Model):
+class HealthDetails(Model):
+    """Health Details for backup items.
+
+    Variables are only populated by the server, and will be ignored when
+    sending a request.
+
+    :ivar code: Health Code
+    :vartype code: int
+    :ivar title: Health Title
+    :vartype title: str
+    :ivar message: Health Message
+    :vartype message: str
+    :ivar recommendations: Health Recommended Actions
+    :vartype recommendations: list[str]
+    """
+
+    _validation = {
+        'code': {'readonly': True},
+        'title': {'readonly': True},
+        'message': {'readonly': True},
+        'recommendations': {'readonly': True},
+    }
+
+    _attribute_map = {
+        'code': {'key': 'code', 'type': 'int'},
+        'title': {'key': 'title', 'type': 'str'},
+        'message': {'key': 'message', 'type': 'str'},
+        'recommendations': {'key': 'recommendations', 'type': '[str]'},
+    }
+
+    def __init__(self, **kwargs):
+        super(HealthDetails, self).__init__(**kwargs)
+        self.code = None
+        self.title = None
+        self.message = None
+        self.recommendations = None
+
+
+class AzureIaaSVMHealthDetails(HealthDetails):
     """Azure IaaS VM workload-specific Health Details.
 
     Variables are only populated by the server, and will be ignored when
@@ -1914,10 +1964,6 @@ class AzureIaaSVMHealthDetails(Model):
 
     def __init__(self, **kwargs):
         super(AzureIaaSVMHealthDetails, self).__init__(**kwargs)
-        self.code = None
-        self.title = None
-        self.message = None
-        self.recommendations = None
 
 
 class Job(Model):
@@ -3460,6 +3506,15 @@ class AzureVmWorkloadProtectedItem(ProtectedItem):
      'Invalid', 'Healthy', 'Unhealthy', 'NotReachable', 'IRPending'
     :type protected_item_health_status: str or
      ~azure.mgmt.recoveryservicesbackup.models.ProtectedItemHealthStatus
+    :param health_status: Health status of protected item. Possible values
+     include: 'Passed', 'ActionRequired', 'ActionSuggested', 'Healthy',
+     'TransientDegraded', 'PersistentDegraded', 'TransientUnhealthy',
+     'PersistentUnhealthy', 'Invalid'
+    :type health_status: str or
+     ~azure.mgmt.recoveryservicesbackup.models.HealthStatus
+    :param health_details: Health details on this backup item.
+    :type health_details:
+     list[~azure.mgmt.recoveryservicesbackup.models.HealthDetails]
     :param extended_info: Additional information for this backup item.
     :type extended_info:
      ~azure.mgmt.recoveryservicesbackup.models.AzureVmWorkloadProtectedItemExtendedInfo
@@ -3495,6 +3550,8 @@ class AzureVmWorkloadProtectedItem(ProtectedItem):
         'last_backup_error_detail': {'key': 'lastBackupErrorDetail', 'type': 'ErrorDetail'},
         'protected_item_data_source_id': {'key': 'protectedItemDataSourceId', 'type': 'str'},
         'protected_item_health_status': {'key': 'protectedItemHealthStatus', 'type': 'str'},
+        'health_status': {'key': 'healthStatus', 'type': 'str'},
+        'health_details': {'key': 'healthDetails', 'type': '[HealthDetails]'},
         'extended_info': {'key': 'extendedInfo', 'type': 'AzureVmWorkloadProtectedItemExtendedInfo'},
     }
 
@@ -3515,6 +3572,8 @@ class AzureVmWorkloadProtectedItem(ProtectedItem):
         self.last_backup_error_detail = kwargs.get('last_backup_error_detail', None)
         self.protected_item_data_source_id = kwargs.get('protected_item_data_source_id', None)
         self.protected_item_health_status = kwargs.get('protected_item_health_status', None)
+        self.health_status = kwargs.get('health_status', None)
+        self.health_details = kwargs.get('health_details', None)
         self.extended_info = kwargs.get('extended_info', None)
         self.protected_item_type = 'AzureVmWorkloadProtectedItem'
 
@@ -3681,6 +3740,15 @@ class AzureVmWorkloadSAPAseDatabaseProtectedItem(AzureVmWorkloadProtectedItem):
      'Invalid', 'Healthy', 'Unhealthy', 'NotReachable', 'IRPending'
     :type protected_item_health_status: str or
      ~azure.mgmt.recoveryservicesbackup.models.ProtectedItemHealthStatus
+    :param health_status: Health status of protected item. Possible values
+     include: 'Passed', 'ActionRequired', 'ActionSuggested', 'Healthy',
+     'TransientDegraded', 'PersistentDegraded', 'TransientUnhealthy',
+     'PersistentUnhealthy', 'Invalid'
+    :type health_status: str or
+     ~azure.mgmt.recoveryservicesbackup.models.HealthStatus
+    :param health_details: Health details on this backup item.
+    :type health_details:
+     list[~azure.mgmt.recoveryservicesbackup.models.HealthDetails]
     :param extended_info: Additional information for this backup item.
     :type extended_info:
      ~azure.mgmt.recoveryservicesbackup.models.AzureVmWorkloadProtectedItemExtendedInfo
@@ -3716,6 +3784,8 @@ class AzureVmWorkloadSAPAseDatabaseProtectedItem(AzureVmWorkloadProtectedItem):
         'last_backup_error_detail': {'key': 'lastBackupErrorDetail', 'type': 'ErrorDetail'},
         'protected_item_data_source_id': {'key': 'protectedItemDataSourceId', 'type': 'str'},
         'protected_item_health_status': {'key': 'protectedItemHealthStatus', 'type': 'str'},
+        'health_status': {'key': 'healthStatus', 'type': 'str'},
+        'health_details': {'key': 'healthDetails', 'type': '[HealthDetails]'},
         'extended_info': {'key': 'extendedInfo', 'type': 'AzureVmWorkloadProtectedItemExtendedInfo'},
     }
 
@@ -4059,6 +4129,15 @@ class AzureVmWorkloadSAPHanaDatabaseProtectedItem(AzureVmWorkloadProtectedItem):
      'Invalid', 'Healthy', 'Unhealthy', 'NotReachable', 'IRPending'
     :type protected_item_health_status: str or
      ~azure.mgmt.recoveryservicesbackup.models.ProtectedItemHealthStatus
+    :param health_status: Health status of protected item. Possible values
+     include: 'Passed', 'ActionRequired', 'ActionSuggested', 'Healthy',
+     'TransientDegraded', 'PersistentDegraded', 'TransientUnhealthy',
+     'PersistentUnhealthy', 'Invalid'
+    :type health_status: str or
+     ~azure.mgmt.recoveryservicesbackup.models.HealthStatus
+    :param health_details: Health details on this backup item.
+    :type health_details:
+     list[~azure.mgmt.recoveryservicesbackup.models.HealthDetails]
     :param extended_info: Additional information for this backup item.
     :type extended_info:
      ~azure.mgmt.recoveryservicesbackup.models.AzureVmWorkloadProtectedItemExtendedInfo
@@ -4094,6 +4173,8 @@ class AzureVmWorkloadSAPHanaDatabaseProtectedItem(AzureVmWorkloadProtectedItem):
         'last_backup_error_detail': {'key': 'lastBackupErrorDetail', 'type': 'ErrorDetail'},
         'protected_item_data_source_id': {'key': 'protectedItemDataSourceId', 'type': 'str'},
         'protected_item_health_status': {'key': 'protectedItemHealthStatus', 'type': 'str'},
+        'health_status': {'key': 'healthStatus', 'type': 'str'},
+        'health_details': {'key': 'healthDetails', 'type': '[HealthDetails]'},
         'extended_info': {'key': 'extendedInfo', 'type': 'AzureVmWorkloadProtectedItemExtendedInfo'},
     }
 
@@ -4507,6 +4588,15 @@ class AzureVmWorkloadSQLDatabaseProtectedItem(AzureVmWorkloadProtectedItem):
      'Invalid', 'Healthy', 'Unhealthy', 'NotReachable', 'IRPending'
     :type protected_item_health_status: str or
      ~azure.mgmt.recoveryservicesbackup.models.ProtectedItemHealthStatus
+    :param health_status: Health status of protected item. Possible values
+     include: 'Passed', 'ActionRequired', 'ActionSuggested', 'Healthy',
+     'TransientDegraded', 'PersistentDegraded', 'TransientUnhealthy',
+     'PersistentUnhealthy', 'Invalid'
+    :type health_status: str or
+     ~azure.mgmt.recoveryservicesbackup.models.HealthStatus
+    :param health_details: Health details on this backup item.
+    :type health_details:
+     list[~azure.mgmt.recoveryservicesbackup.models.HealthDetails]
     :param extended_info: Additional information for this backup item.
     :type extended_info:
      ~azure.mgmt.recoveryservicesbackup.models.AzureVmWorkloadProtectedItemExtendedInfo
@@ -4542,6 +4632,8 @@ class AzureVmWorkloadSQLDatabaseProtectedItem(AzureVmWorkloadProtectedItem):
         'last_backup_error_detail': {'key': 'lastBackupErrorDetail', 'type': 'ErrorDetail'},
         'protected_item_data_source_id': {'key': 'protectedItemDataSourceId', 'type': 'str'},
         'protected_item_health_status': {'key': 'protectedItemHealthStatus', 'type': 'str'},
+        'health_status': {'key': 'healthStatus', 'type': 'str'},
+        'health_details': {'key': 'healthDetails', 'type': '[HealthDetails]'},
         'extended_info': {'key': 'extendedInfo', 'type': 'AzureVmWorkloadProtectedItemExtendedInfo'},
     }